Moved to DynamicMapLayer instead of SingleTIle Functionality is equal. Also wms-esri layer made like wms-wfs layer, using inner esri-dynamic
