+++ addon/components/layers/wms-esri-layer.js
@@ -1,222 +1,113 @@
 import Ember from 'ember';
-import WMSLayer from 'ember-flexberry-gis/components/layers/wms-layer';
+import WMSLayerComponent from 'ember-flexberry-gis/components/layers/wms-layer';
+import EsriDynamycLayerComponent from './esri-dynamic-layer';
 
-let projectBounds = function (boundingBox, crs) {
-  let sw = crs.project(boundingBox.getSouthWest());
-  let ne = crs.project(boundingBox.getNorthEast());
+/**
+  WMS Esri component for leaflet map.
 
-  return [
-    sw.x,
-    sw.y,
-    ne.x,
-    ne.y
-  ];
-};
-
-export default WMSLayer.extend({
+  @class WmsEsriLayerComponent
+  @extends WMSLayerComponent
+ */
+export default WMSLayerComponent.extend({
   /**
-    Url of ArcGIS REST service endpoint
+    Inner ESRI dynamic layer.
+    Needed for identification (always invisible, won't be added to map).
+
+    @property _esriDynamicLayer
+  */
+  _esriDynamicLayer: null,
+
+  /**
+     Handles 'flexberry-map:identify' event of leaflet map.
+
+     @method identify
+     @param {Object} e Event object.
+     @param {<a href="http://leafletjs.com/reference-1.0.0.html#latlngbounds">L.LatLngBounds</a>} options.boundingBox Bounds of identification area.
+     @param {<a href="http://leafletjs.com/reference-1.0.0.html#latlng">L.LatLng</a>} e.latlng Center of the bounding box.
+     @param {Object[]} layers Objects describing those layers which must be identified.
+     @param {Object[]} results Objects describing identification results.
+     Every result-object has the following structure: { layer: ..., features: [...] },
+     where 'layer' is metadata of layer related to identification result, features is array
+     containing (GeoJSON feature-objects)[http://geojson.org/geojson-spec.html#feature-objects]
+     or a promise returning such array.
    */
-  restUrl: null,
-
-  layersIds: Ember.computed('restUrl', function () {
-    let mapLayers = this.get('layerModel.layers');
-
-    if (Ember.isArray(mapLayers) && mapLayers.length !== 0) {
-      return mapLayers.map(layer => layer.id);
-    } else {
-      let params = {
-        f: 'json'
-      };
-
-      let url = this.get('restUrl');
-      url = url + L.Util.getParamString(params, url);
-
-      return new Ember.RSVP.Promise((resolve, reject) => {
-
-        Ember.$.ajax({
-          url,
-          success: function (data) {
-            let result = JSON.parse(data);
-            resolve(result.layers.filter(layer => !layer.subLayerIds).map(layer => layer.id));
-          },
-          error: function () {
-            reject();
-          }
-        });
-      });
-    }
-  }),
-
-  _queryLayer(layerId, params) {
-    params = Ember.$.extend(true, params, {
-      f: 'json'
-    });
-    return new Ember.RSVP.Promise((resolve, reject) => {
-      let url = this.get('restUrl') + '/' + layerId + '/query';
-      let crs = this.get('crs');
-      url = url + L.Util.getParamString(params, url);
-
-      Ember.$.ajax({
-        url,
-        success: function (data) {
-          let result = JSON.parse(data);
-          if (result.error) {
-            reject(result.error);
-            return;
-          }
-
-          let featureCollection = L.esri.Util.responseToFeatureCollection(result);
-          let coordsToLatLng = function (coords) {
-            var point = L.point(coords[0], coords[1]);
-            var ll = crs.projection.unproject(point);
-            if (coords[2]) {
-              ll.alt = coords[2];
-            }
-
-            return ll;
-          };
-
-          resolve(L.geoJSON(featureCollection, {
-            coordsToLatLng
-          }));
-        },
-        error: function (e) {
-          reject(e);
-        }
-      });
-    });
-  },
-
-  _query(e) {
-    let layerId = this.get('layerModel.id').toString();
-    let layerLinks = this.get('layerModel.layerLink');
-    if (!layerLinks.length) {
-      return;
-    }
-
-    let queryFilter = e.queryFilter;
-
-    e.results.push(new Ember.RSVP.Promise((resolve, reject) => {
-      let result = L.featureGroup();
-
-      this.get('layersIds')
-        .then(layerIds => {
-          let allQueries = [];
-          layerIds.forEach((layerId) => {
-            layerLinks.forEach((link) => {
-              let params = {
-                where: link.get('linkParameter').map(link => link.get('layerField') + '=' + queryFilter[link.get('queryKey')]).join(' and ')
-              };
-
-              allQueries.push(this._queryLayer(layerId, params));
-            });
-          });
-
-          Ember.RSVP.all(allQueries).then(layers => {
-            layers.forEach(layer => result.addLayer(layer));
-            resolve(result);
-          });
-        })
-        .catch((error) => {
-          reject(error);
-        });
-    }));
-  },
-
-  didInsertElement() {
-    this._super(...arguments);
-    this.get('leafletMap').on('flexberry-map:query', this._query, this);
-  },
-
-  _identifyEsri(boundingBox) {
-    return new Ember.RSVP.Promise((resolve, reject) => {
-      let url = this.get('restUrl') + '/identify';
-      let crs = this.get('crs');
-      let map = this.get('leafletMap');
-      let size = map.getSize();
-
-      let params = {
-        f: 'json',
-        geometryType: 'esriGeometryEnvelope',
-        layers: 'visible',
-        returnGeometry: true,
-        tolerance: 5,
-        geometry: projectBounds(boundingBox, crs),
-        imageDisplay: [size.x, size.y, 96],
-        mapExtent: projectBounds(map.getBounds(), crs)
-      };
-
-      url = url + L.Util.getParamString(params, url);
-
-      let _this = this;
-
-      Ember.$.ajax({
-        url,
-        success: function (data) {
-          let result = JSON.parse(data);
-          let featureCollection = L.esri.Util.responseToFeatureCollection(result);
-          _this.injectLeafletLayersIntoGeoJSON(featureCollection);
-          resolve(Ember.A(Ember.get(featureCollection, 'features') || []));
-        },
-        error: function (e) {
-          reject(e);
-        }
-      });
-    });
-  },
-
   identify(e) {
-    let featuresPromise = this._identifyEsri(e.boundingBox);
-    e.results.push({
-      layer: this.get('layerModel'),
-      features: featuresPromise
-    });
+    let innerEsriDynamicLayer = this.get('_esriDynamicLayer');
+    if (!Ember.isNone(innerEsriDynamicLayer)) {
+      innerEsriDynamicLayer.identify.apply(innerEsriDynamicLayer, arguments);
+    }
   },
 
+  query(e) {
+    let innerEsriDynamicLayer = this.get('_esriDynamicLayer');
+    if (!Ember.isNone(innerEsriDynamicLayer)) {
+      innerEsriDynamicLayer.query.apply(innerEsriDynamicLayer, arguments);
+    }
+  },
+
+  /**
+    Handles 'flexberry-map:search' event of leaflet map.
+
+    @method search
+    @param {Object} e Event object.
+    @param {<a href="http://leafletjs.com/reference-1.0.0.html#latlng">L.LatLng</a>} e.latlng Center of the search area.
+    @param {Object[]} layer Object describing layer that must be searched.
+    @param {Object} searchOptions Search options related to layer type.
+    @param {Object} results Hash containing search results.
+    @param {Object[]} results.features Array containing (GeoJSON feature-objects)[http://geojson.org/geojson-spec.html#feature-objects]
+    or a promise returning such array.
+  */
   search(e) {
-    return new Ember.RSVP.Promise((resolve, reject) => {
-      this.get('layersIds')
-        .then(layerIds => {
-          let allQueries = [];
-          layerIds.forEach((layerId) => {
-            let queryProperties = {
-              outFields: '*'
-            };
-            let searchFields = this.get('searchSettings.searchFields');
-            if (Ember.isNone(searchFields)) {
-              queryProperties.text = e.searchOptions.queryString.replace(/ /g, '%');
-            } else {
-              let whereClause = [];
-              searchFields.split(',').forEach(function (field) {
-                whereClause.push(field + ' like \'%' + e.searchOptions.queryString.replace(/ /g, '%') + '%\'');
-              });
+    let innerEsriDynamicLayer = this.get('_esriDynamicLayer');
+    if (!Ember.isNone(innerEsriDynamicLayer)) {
+      return innerEsriDynamicLayer.search.apply(innerEsriDynamicLayer, arguments);
+    }
+  },
 
-              queryProperties.where = whereClause.join(' OR ');
-            }
+  /**
+    Initializes component.
+  */
+  init() {
+    this._super(...arguments);
 
-            allQueries.push(this._queryLayer(layerId, queryProperties));
-          });
+    let innerEsriDynamicLayerProperties = {
+      leafletMap: this.get('leafletMap'),
+      leafletContainer: this.get('leafletContainer'),
+      layerModel: this.get('layerModel'),
+      crs: this.get('crs'),
+      index: this.get('index'),
+      visibility: this.get('visibility'),
+      restUrl: this.get('restUrl'),
+      esriLayers: this.get('esriLayers')
+    };
 
-          Ember.RSVP.all(allQueries).then(featureLayers => {
-              let features = Ember.A();
-              featureLayers.forEach(featureLayer => {
-                featureLayer.eachLayer(layer => {
-                  let feature = layer.feature;
-                  feature.leafletLayer = layer;
-                  features.push(feature);
-                });
-              });
-
-              resolve(features);
-            },
-            (error) => {
-              reject(error);
-            });
-        })
-        .catch((error) => {
-          reject(error);
-        });
+    // Set creating component's owner to avoid possible lookup exceptions.
+    let owner = Ember.getOwner(this);
+    let ownerKey = null;
+    Ember.A(Object.keys(this) || []).forEach((key) => {
+      if (this[key] === owner) {
+        ownerKey = key;
+        return false;
+      }
     });
+    if (!Ember.isBlank(ownerKey)) {
+      innerEsriDynamicLayerProperties[ownerKey] = owner;
+    }
+
+    // Create inner ESRI dynamic layer which is needed for identification (always invisible, won't be added to map).
+    this.set('_esriDynamicLayer', EsriDynamycLayerComponent.create(innerEsriDynamicLayerProperties));
+  },
+
+  /**
+    Destroys component.
+  */
+  willDestroyElement() {
+    this._super(...arguments);
+
+    let innerEsriDynamicLayer = this.get('_esriDynamicLayer');
+    if (!Ember.isNone(innerEsriDynamicLayer)) {
+      innerEsriDynamicLayer.destroy();
+      this.set('_esriDynamicLayer', null);
+    }
   }
 });
