--- addon/layers/wms-single-tile-esri.js
@@ -1,6 +1,0 @@
-/**
-  Class describing wms single tile layer metadata.
-
-  @class WmsSingleTileLayer
-*/
-export { default } from './wms-esri';
