+++ addon/templates/components/map-commands-dialogs/search-settings/esri-dynamic.hbs
