+++ app/components/layers/esri-dynamic.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-flexberry-gis-esri/components/layers/esri-dynamic-layer';
