--- app/components/layers/wms-single-tile-esri-layer.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-flexberry-gis-esri/components/layers/wms-single-tile-esri-layer';
