+++ app/components/map-commands-dialogs/search-settings/esri-dynamic.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-flexberry-gis-esri/components/map-commands-dialogs/search-settings/esri-dynamic';
