--- app/components/map-commands-dialogs/search-settings/wms-single-tile-esri.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-flexberry-gis-esri/components/map-commands-dialogs/search-settings/wms-single-tile-esri';
