+++ app/layers/esri-dynamic.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-flexberry-gis-esri/layers/esri-dynamic';
