--- app/layers/wms-single-tile-esri.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-flexberry-gis-esri/layers/wms-single-tile-esri';
